elk_nimble/kafka-adapter/code/requirements.txt
python-logstash
confluent-kafka
flask
redis
